add Arithmetic trait
