+++ src/machine/mod.rs
@@ -1,7 +1,11 @@
-use crate::lir::Term;
+use crate::{
+    lir::Term,
+    machine::arithmetic::{Arithmetic, OverflowArithmetic},
+};
 
 use std::io::{stdout, Stdout, Write};
 
+pub mod arithmetic;
 mod eval;
 
 pub struct LangEnv<W: Write> {
@@ -14,25 +18,30 @@ impl Default for LangEnv<Stdout> {
     }
 }
 
-pub struct Machine<W: Write> {
+pub struct Machine<W: Write, A: Arithmetic> {
     env: LangEnv<W>,
+    arithmetic: A,
 }
 
-impl Default for Machine<Stdout> {
+impl Default for Machine<Stdout, OverflowArithmetic> {
     fn default() -> Self {
         Machine {
             env: LangEnv::default(),
+            arithmetic: OverflowArithmetic,
         }
     }
 }
 
-impl<W: Write> Machine<W> {
+impl<W: Write> Machine<W, OverflowArithmetic> {
     pub fn with_env(env: LangEnv<W>) -> Self {
-        Machine { env }
+        Machine {
+            env,
+            arithmetic: OverflowArithmetic,
+        }
     }
 }
 
-impl<W: Write> Machine<W> {
+impl<W: Write, A: Arithmetic> Machine<W, A> {
     pub fn evaluate(&mut self, mut term: Term) -> Term {
         while {
             let (eval, new_term) = self.step(term);
